Test shorter periods
Down to about a 6us period, it seems it's pretty stable (according to the scope); below that and we have noticeable slowdown and jitter. This is all to be expected as we're still just driving GPIO naively from one of the CPU cores. I'd expect using one of the PIO state machines as a FIFO would smooth this out (more or less; assuming we can keep up with the data rate at least).