Working DAC test
So far just a simple ramp and it's fairly noisy on ths scope (both because of actual noise, esp. on a breadboard with long wires, as well as a lazy probing setup) but it's more than good enough for now.

The DAC design follows this sim exactly: https://everycircuit.com/circuit/6379524121690112/rpi-pico-video-dac-sim . The current test rig even includes an extra 75ohm load to model the input impedance of the receiver since that's currently not connected.